Refactor Login frontend component
